Refactor transport specific timeout into separate module
This fixes pylint complaining about circular imports.
