+++ adbts/usb/__init__.py
@@ -4,15 +4,6 @@
 
     Package that contains Universal Serial Bus (USB) transports.
 """
-import functools
-
-from .. import timeouts
-
-
-#: Function partial that sets the default USB transport operation timeout to zero milliseconds.
-usb_timeout = functools.partial(timeouts.timeout, default=0)  # pylint: disable=invalid-name
-
-
-from . import asynchronous, synchronous  # noqa pylint: disable=wrong-import-position
+from . import asynchronous, synchronous
 
 __all__ = ['asynchronous', 'synchronous']
